Standardize integer types

+++ src/attack_tables/bishop_attack_generators.rs
@@ -3,8 +3,8 @@ use crate::utils::{board_slice::BoardSlice, enums::Square};
 pub fn generate_bishop_attack_mask(square: Square) -> BoardSlice {
     let mut attack_mask = BoardSlice(0);
 
-    let rank = square as i64 / 8;
-    let file = square as i64 % 8;
+    let rank = square as usize / 8;
+    let file = square as usize % 8;
 
     for (i, j) in ((rank + 1)..7).zip((file + 1)..7) {
         attack_mask.0 |= 1 << (i * 8 + j);
@@ -25,8 +25,8 @@ pub fn generate_bishop_attack_mask(square: Square) -> BoardSlice {
 pub fn generate_bishop_attacks_on_the_fly(square: Square, blockers: BoardSlice) -> BoardSlice {
     let mut attacks = BoardSlice(0);
 
-    let rank = square as i64 / 8;
-    let file = square as i64 % 8;
+    let rank = square as usize / 8;
+    let file = square as usize % 8;
 
     for (i, j) in ((rank + 1)..8).zip((file + 1)..8) {
         attacks.0 |= 1 << (i * 8 + j);
@@ -85,7 +85,9 @@ pub mod tests {
             generate_bishop_attacks_on_the_fly(
                 Square::D4,
                 BoardSlice(
-                    (1 << Square::B2 as u64) | (1 << Square::E3 as u64) | (1 << Square::F6 as u64)
+                    (1 << Square::B2 as usize)
+                        | (1 << Square::E3 as usize)
+                        | (1 << Square::F6 as usize)
                 )
             ),
             BoardSlice(0x1221400140200)
@@ -95,7 +97,9 @@ pub mod tests {
             generate_bishop_attacks_on_the_fly(
                 Square::F3,
                 BoardSlice(
-                    (1 << Square::E2 as u64) | (1 << Square::G2 as u64) | (1 << Square::G4 as u64)
+                    (1 << Square::E2 as usize)
+                        | (1 << Square::G2 as usize)
+                        | (1 << Square::G4 as usize)
                 )
             ),
             BoardSlice(0x102040850005000)
